bot/requirements.txt
python-telegram-bot==21.10
python-dotenv==1.1.0
httpx==0.28.1
mcp==1.9.2
duckduckgo-search==7.5.1
